# openapi.yaml
openapi: 3.0.3
info:
  title: Young Artist Community API
  description: |
    REST API for Young Artist Community, a directory and review platform for
    Young Artist Programs (YAPs) in classical music and opera.
  version: 0.2.0
  contact:
    name: Young Artist Community
    email: api@youngartist.community
    url: https://youngartist.community
  x-api-id: young-artist-community-api
  x-audience: external-public
servers:
  - url: https://youngartist.community

# ---------------------------------------------------------------------------
# PATHS
# ---------------------------------------------------------------------------
paths:
...
  /programs/{program_id}/reviews:
    get:
      summary: List reviews for a program
      operationId: listProgramReviews
      parameters:
        - $ref: '#/components/parameters/ProgramId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - name: rating
          in: query
          description: Filter by exact rating (1-5).
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: year_attended
          in: query
          description: Filter by the year the reviewer attended.
          schema:
            type: integer
      responses:
        '200':
          description: A cursor-paginated list of reviews for the program.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultError'

    post:
      summary: Create a review for a program
      operationId: createReview
      parameters:
        - $ref: '#/components/parameters/ProgramId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '201':
          description: Review created successfully.
          headers:
            Location:
              description: URL of the newly created review.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultError'
  /programs/{program_id}/reviews/{review_id}:
    get:
      summary: Get a single review
      operationId: getReview
      parameters:
        - $ref: '#/components/parameters/ProgramId'
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '200':
          description: The requested review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultError'

    put:
      summary: Update a review
      operationId: updateReview
      parameters:
        - $ref: '#/components/parameters/ProgramId'
        - $ref: '#/components/parameters/ReviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '200':
          description: Review updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultError'

  # ---- Auditions (sub-resource of Program) --------------------------------
...
  /locations:
    get:
      summary: List all locations
      operationId: listLocations
      parameters:
        - name: country
          in: query
          description: Filter locations by country.
          schema:
            type: string
      responses:
        '200':
          description: List of all locations.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultError'

    post:
      summary: Create a location
      operationId: createLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationInput'
      responses:
        '201':
          description: Location created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultError'
...
components:
...
  parameters:
    ProgramId:
      name: program_id
      in: path
      required: true
      description: UUID of the program.
      schema:
        type: string
        format: uuid
    ReviewId:
      name: review_id
      in: path
      required: true
      description: UUID of the review.
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: |
        Opaque cursor token identifying the start of the page. Obtain from
        the `next` or `prev` field of a previous list response. Omit to
        fetch the first page.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maximum number of items per page.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      description: |
        Comma-separated list of fields to sort by. Prefix with `-` for
        descending order. Example: `-created_at,name`
      schema:
        type: string

  # ---- Reusable responses -------------------------------------------------
  responses:
    BadRequest:
      description: Malformed request or invalid parameters.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError:
      description: Unexpected server error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    DefaultError:
      description: Unexpected error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

  # ---- Schemas ------------------------------------------------------------
  schemas:
...
    Problem:
      type: object
      description: Problem detail object per RFC 9457.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: |
            A URI reference that identifies the problem type. When dereferenced,
            it SHOULD provide human-readable documentation for the problem type.
          example: /problems/not-found
        title:
          type: string
          description: A short human-readable summary of the problem type.
          example: Resource not found
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 600
          exclusiveMaximum: true
          description: The HTTP status code generated for this occurrence.
          example: 404
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
          example: Program with id '550e8400-e29b-41d4-a716-446655440000' not found.
        instance:
          type: string
          format: uri-reference
          description: A URI reference that identifies this specific occurrence of the problem.

    # -- Cursor pagination meta --
    CursorPageMeta:
      type: object
      description: |
        Cursor-based pagination metadata. `next` and `prev` are opaque tokens
        passed to the `cursor` query parameter on subsequent requests. Both
        are null when no further page exists in that direction.
      properties:
        next:
          type: string
          nullable: true
          description: Opaque cursor for the next page, or null if this is the last page.
        prev:
          type: string
          nullable: true
          description: Opaque cursor for the previous page, or null if this is the first page.
        total_items:
          type: integer
          format: int32
          description: Total number of items matching the query across all pages.

    # -- Program --
...
    Review:
      type: object
      required:
        - id
        - program_id
        - rating
        - body
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        program_id:
          type: string
          format: uuid
          readOnly: true
        rating:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        year_attended:
          type: integer
          format: int32
          example: 2025
        reviewer_name:
          type: string
          description: Display name of the reviewer (anonymous if omitted).
          example: Soprano2025
        title:
          type: string
          example: Transformative experience
        body:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    ReviewInput:
      type: object
      required:
        - rating
        - body
      properties:
        rating:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        year_attended:
          type: integer
          format: int32
        reviewer_name:
          type: string
        title:
          type: string
        body:
          type: string
    ReviewListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        meta:
          $ref: '#/components/schemas/CursorPageMeta'

    # -- Audition --
...
    LocationInput:
      type: object
      required:
        - city
        - country
      properties:
        city:
          type: string
        country:
          type: string
        state:
          type: string
        address:
          type: string

    # -- Reference data --
...
    Location:
      type: object
      required:
        - id
        - city
        - country
      properties:
        id:
          type: string
          format: uuid
        city:
          type: string
          example: Salzburg
        country:
          type: string
          example: Austria
        state:
          type: string
        address:
          type: string